+++ game/levelable.h
@@ -13,25 +13,25 @@ namespace game {
 		virtual void setExperience(int experience);
 		virtual int getExperience() const;
 
-		virtual void addLevel(int level);
-		virtual void setLevel(int level);
-		virtual int getLevel() const;
+		virtual void addLevel(short level);
+		virtual void setLevel(short level);
+		virtual short getLevel() const;
 
-		virtual void setMaxLevel(int level);
-		virtual int getMaxLevel() const;
+		virtual void setMaxLevel(short level);
+		virtual short getMaxLevel() const;
 
-		virtual int getExperienceForLevel(int level) const;
-		virtual int getLevelFromExperience(int experience) const;
+		virtual int getExperienceForLevel(short level) const;
+		virtual short getLevelFromExperience(int experience) const;
 
 	protected:
 
-		int mLevel;
-		int mMaxLevel;
+		short mLevel;
+		short mMaxLevel;
 		int mExperience;
 
 		virtual void onLevelUp() {}
 
-		virtual void addLevels(int level);
+		virtual void addLevels(short level);
 
 	};
